requirements.txt
bootstrap-admin==0.3.7.1
simplejson==3.11.1
Django==2.0
django-environ==0.4.3
django-filter==1.0.4
djangorestframework==3.7.1
djangorestframework-filters==0.8.0
djangorestframework-jwt==1.11.0
django-simple-history==1.8.2
django-reversion==2.0.8
psycopg2==2.7.3
python-dateutil==2.6.1
requests==2.18.1
beautifulsoup4==4.6.0
retrying==1.3.3
redis==2.10.5
django-redis==4.8.0
isodate==0.5.4
djoser==0.6.0
pillow==4.2.1
